float sumOf

+++ src/com/genir/aitweaks/core/features/shipai/ShipStats.kt
@@ -5,6 +5,7 @@ import com.fs.starfarer.api.combat.WeaponAPI
 import com.genir.aitweaks.core.utils.extensions.isAngleInArc
 import com.genir.aitweaks.core.utils.extensions.isInFiringSequence
 import com.genir.aitweaks.core.utils.extensions.isPDSpec
+import com.genir.aitweaks.core.utils.extensions.sumOf
 import com.genir.aitweaks.core.utils.shortestRotation
 import kotlin.math.abs
 import kotlin.math.max
@@ -86,8 +87,8 @@ class ShipStats(private val ship: ShipAPI) {
 
         // Calculate DPS for each attack angle.
         val anglesDPS: List<AngleDPS> = attackAngles.map { angle ->
-            val dps = significantWeapons.filter { it.isAngleInArc(angle) }.sumOf { it.derivedStats.dps.toDouble() }
-            AngleDPS(angle, dps.toFloat())
+            val dps = significantWeapons.filter { it.isAngleInArc(angle) }.sumOf { it.derivedStats.dps }
+            AngleDPS(angle, dps)
         }
 
         // Find all weapon groups with acceptable DPS.
